+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changelog
 - `aruco-detector` stores the detections in the annotations as well, in case the data passing through
   is of the object detection data domain
 - fixed margin handling in `idc-generate-regions` tool when not using fixed sizes
+- added `overlay-regions` filter to visualize, e.g., regions used by `sub-images` or `meta-sub-images`
 
 
 0.1.0 (2025-10-31)